requirements.txt
flake8<7
pytest-cov<5

-e .
